transport: move the public cred structure somewhere less visible
